+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Added a generic JSON parser that reads all properties of the internal `Issue` object.
 - [PR#57](https://github.com/jenkinsci/warnings-ng-plugin/pull/57):
 Added a parser for MentorGraphcis Modelsim/Questa.
+- [PR#111](https://github.com/jenkinsci/warnings-ng-plugin/pull/111):
+  Added a Kotlin Parser.
 
 ### Fixed 
 - [JENKINS-56007](https://issues.jenkins-ci.org/browse/JENKINS-56007): Obtain the affected files in a process on the